refactor(folder): api 관련 파일 수정
